fix: code review changes
