+++ src/modals/new-identity-flow/create-identity-steps/identity-device/IdentityDevice.css
@@ -1,0 +1,44 @@
+@import "@nomios/web-uikit/styles/colors";
+
+.contentWrapper {
+    width: 100%;
+    height: 100%;
+    padding: 3rem 6rem;
+    display: flex;
+    flex-direction: column;
+
+    & .heading {
+        margin-top: 0;
+        margin-bottom: 3rem;
+        color: var(--color-tan);
+    }
+
+    & > p {
+        max-width: 65rem;
+        margin-bottom: 8.4rem;
+        opacity: 0.7;
+        color: var(--color-armadillo);
+    }
+
+    & .buttonWrapper {
+        margin-top: auto;
+        align-self: flex-end;
+    }
+
+    & .typeGroup {
+        & > *:not(:last-child) {
+            margin-right: 9.5rem;
+        }
+    }
+
+    & .deviceInfoWrapper {
+        display: flex;
+        align-items: center;
+
+        & .textInput {
+            width: 40.6rem;
+            margin-left: 6.8rem;
+            transform: translateY(-25%);
+        }
+    }
+}
