+++ src/modals/new-identity-flow/create-identity-steps/identity-device/IdentityDevice.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import SetupDeviceStep from '../../../../shared/steps/setup-device-step';
+
+const stepData = {
+    title: '2. Setup your device',
+    description: 'To complete your identity creation you must enter your device name and select which type of device are you on.',
+    buttonText: 'Continue',
+};
+
+const IdentityDevice = (props) => <SetupDeviceStep { ...props } stepData={ stepData } />;
+
+export default IdentityDevice;
