+++ src/modals/new-identity-flow/create-identity-steps/identity-device/index.js
@@ -1,0 +1,3 @@
+import IdentityDevice from './IdentityDevice';
+
+export default IdentityDevice;
