+++ src/modals/new-identity-flow/create-identity-steps/identity-info/IdentityInfo.css
@@ -1,6 +1,7 @@
 @import "@nomios/web-uikit/styles/colors";
 
 .contentWrapper {
+    position: relative;
     width: 100%;
     height: 100%;
     padding: 3rem 6rem;
@@ -20,25 +21,36 @@
         color: var(--color-armadillo);
     }
 
-    & .buttonWrapper {
-        margin-top: auto;
-        align-self: flex-end;
-    }
-
-    & .typeGroup {
-        & > *:not(:last-child) {
-            margin-right: 9.5rem;
-        }
-    }
-
-    & .deviceInfoWrapper {
+    & form {
         display: flex;
-        align-items: center;
+        flex-grow: 1;
+        flex-direction: column;
 
-        & .textInput {
-            width: 40.6rem;
-            margin-left: 6.8rem;
-            transform: translateY(-25%);
+        & .buttonWrapper {
+            margin-top: auto;
+            align-self: flex-end;
         }
+
+        & .identityInfoWrapper {
+            display: flex;
+            align-items: center;
+
+            & .textInput {
+                width: 40.6rem;
+                margin-left: 6.8rem;
+                transform: translateY(-25%);
+
+                & input {
+                    background-color: transparent;
+                }
+            }
+        }
+    }
+
+    & .bulletsIndicator {
+        position: absolute;
+        top: 50%;
+        right: 6rem;
+        transform: translateY(-50%);
     }
 }
