--- src/modals/new-identity-flow/create-identity-steps/identity-type/IdentityType.css
@@ -1,56 +1,0 @@
-@import "@nomios/web-uikit/styles/colors";
-
-.contentWrapper {
-    position: relative;
-    width: 100%;
-    height: 100%;
-    padding: 3rem 6rem;
-    display: flex;
-    flex-direction: column;
-
-    & .heading {
-        margin-top: 0;
-        margin-bottom: 3rem;
-        color: var(--color-tan);
-    }
-
-    & > p {
-        max-width: 65rem;
-        margin-bottom: 8.4rem;
-        opacity: 0.7;
-        color: var(--color-armadillo);
-    }
-
-    & form {
-        display: flex;
-        flex-grow: 1;
-        flex-direction: column;
-
-        & .buttonWrapper {
-            margin-top: auto;
-            align-self: flex-end;
-        }
-
-        & .identityInfoWrapper {
-            display: flex;
-            align-items: center;
-
-            & .textInput {
-                width: 40.6rem;
-                margin-left: 6.8rem;
-                transform: translateY(-25%);
-
-                & input {
-                    background-color: transparent;
-                }
-            }
-        }
-    }
-
-    & .bulletsIndicator {
-        position: absolute;
-        top: 50%;
-        right: 6rem;
-        transform: translateY(-50%);
-    }
-}
