--- src/modals/new-identity-flow/create-identity-steps/identity-type/index.js
@@ -1,3 +1,0 @@
-import IdentityType from './IdentityType';
-
-export default IdentityType;
