+++ src/modals/new-identity-flow/create-identity-steps/index.js
@@ -1,3 +1,3 @@
-export { default as IdentityType } from './identity-type';
 export { default as IdentityInfo } from './identity-info';
+export { default as IdentityDevice } from './identity-device';
 export { default as Feedback } from './feedback';
