+++ src/pages/home/Home.js
@@ -1,14 +1,8 @@
 import React from 'react';
-import { ModalTrigger } from '@nomios/web-uikit';
-
-import NewIdentityFlow from '../../modals/new-identity-flow';
 
 const Home = () => (
     <div>
         <h4>This is the home page!</h4>
-        <ModalTrigger modal={ <NewIdentityFlow /> }>
-            <button>Add new identity</button>
-        </ModalTrigger>
     </div>
 );
