+++ src/shared/components/fade-container/FadeContainer.css
