+++ src/shared/components/fade-container/index.js
@@ -1,0 +1,3 @@
+import FadeContainer from './FadeContainer';
+
+export default FadeContainer;
