--- src/shared/components/fader-container/index.js
@@ -1,3 +1,0 @@
-import FaderContainer from './FaderContainer';
-
-export default FaderContainer;
