--- src/shared/utils/index.js
@@ -1,5 +1,0 @@
-import capitalize from './capitalize';
-
-export {
-    capitalize,
-};
