+++ src/shell/Shell.css
@@ -5,7 +5,6 @@
 
     & .sidebar {
         flex: 0 0 auto;
-        background: red;
     }
 
     & .content {
